sql: initialize all members of Response in ctor
Found by cppcheck:
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::name' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::outlist' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::height' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::floatval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::stringval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::linkval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::table' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::store' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::pvec' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::uvec' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::fltval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::strval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::lnkval' is not initialized in the constructor.
[opencog/persist/sql/multi-driver/SQLAtomStorage.cc:96]: (warning) Member
variable 'Response::get_all_values' is not initialized in the constructor.