Avoid confusion due to re-ordering of arguments.
Some JSON tools will alphabetize the fields in a json object, which
wrecks our parsing code which assumed that they can in order. This
sucks, because (1) it makes the code slower, and (2) lose time debuging
this stuff, and (3) growth of complecity and harder-to-maintain code.
So this sucks all around in all ways. But the frigging JSON tools do
this so we have to put up with this bullshit.  I'm pissed because
I've spent maybe six hours trying to fix this.

At any rate, Claude did most of the work here. Its still not right,
but I think its closer to working.