Decouple storage remove from local remove.
Long long ago, in a galaxy far far away, backing store was conceived
of as some storage medium that "backs up" the local, in-RAM database.
In this model, whatever one did localy eventually perfolated there.
Later, the roles wre reversed: the local in-RAM database was a cache
for the remote storage. Either way, removing an Atom from Storage
seemed to imply that it should be removed from the local AtomSpace.

This no longer becomes tenable, when all thos remote nodes are
imagined to be distinct entities. We can manipulate them there,
but are under no obligation to shoot ourselves in the foot here.
The two-stage commit design for removal is simultaneously brilliant
and brain-damaged. Just don't do it. It's too complex, too crazy.
What happens in Vegas stays in Vegas.